Refactor list mixins and enhance test coverage
Refactored `ObjectListMixin` and `GenericListMixin` by splitting them into modular, reusable components under `simstack.util`. Replaced legacy list implementations in models (e.g., `ArrayList`) with streamlined mixins for consistency. Additionally, implemented new test cases to validate list behaviors and ensure robust database interactions in unit tests.
